# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AdhocExecutionRequest": {
        "description": "Run one framework directly, without a cycle in ``testosterone.yaml``.",
        "properties": {
          "args": {
            "items": {
              "type": "string"
            },
            "title": "Args",
            "type": "array"
          },
          "artifacts_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts Root"
          },
          "config_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Path"
          },
          "extra_env": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Env"
          },
          "framework": {
            "enum": [
              "pytest",
              "behave",
              "behavex",
              "command"
            ],
            "title": "Framework",
            "type": "string"
          },
          "persist": {
            "default": true,
            "title": "Persist",
            "type": "boolean"
          },
          "report_db": {
            "default": true,
            "title": "Report Db",
            "type": "boolean"
          },
          "target_repo": {
            "title": "Target Repo",
            "type": "string"
          },
          "timeout_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout S"
          }
        },
        "required": [
          "framework",
          "target_repo"
        ],
        "title": "AdhocExecutionRequest",
        "type": "object"
      },
      "AiConfigStatusResponse": {
        "properties": {
          "api_key_env_var": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Env Var"
          },
          "api_key_source": {
            "enum": [
              "env",
              "runtime_input"
            ],
            "title": "Api Key Source",
            "type": "string"
          },
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "key_present": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key Present"
          },
          "max_input_chars": {
            "title": "Max Input Chars",
            "type": "integer"
          },
          "max_output_tokens": {
            "title": "Max Output Tokens",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "enum": [
              "openai",
              "anthropic"
            ],
            "title": "Provider",
            "type": "string"
          },
          "retry_count": {
            "title": "Retry Count",
            "type": "integer"
          },
          "timeout_s": {
            "title": "Timeout S",
            "type": "number"
          }
        },
        "required": [
          "enabled",
          "configured",
          "provider",
          "model",
          "api_key_source",
          "api_key_env_var",
          "key_present",
          "timeout_s",
          "retry_count",
          "max_input_chars",
          "max_output_tokens"
        ],
        "title": "AiConfigStatusResponse",
        "type": "object"
      },
      "AiConfigUpdateRequest": {
        "properties": {
          "api_key_env_var": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Env Var"
          },
          "api_key_input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Input"
          },
          "api_key_source": {
            "default": "env",
            "enum": [
              "env",
              "runtime_input"
            ],
            "title": "Api Key Source",
            "type": "string"
          },
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "max_input_chars": {
            "default": 16000,
            "title": "Max Input Chars",
            "type": "integer"
          },
          "max_output_tokens": {
            "default": 300,
            "title": "Max Output Tokens",
            "type": "integer"
          },
          "model": {
            "default": "gpt-4o-mini",
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "default": "openai",
            "enum": [
              "openai",
              "anthropic"
            ],
            "title": "Provider",
            "type": "string"
          },
          "retry_count": {
            "default": 1,
            "title": "Retry Count",
            "type": "integer"
          },
          "timeout_s": {
            "default": 8.0,
            "title": "Timeout S",
            "type": "number"
          }
        },
        "title": "AiConfigUpdateRequest",
        "type": "object"
      },
...
      "CycleDetailResponse": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageSummary"
            },
            "title": "Stages",
            "type": "array"
          },
          "trigger": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CycleTriggerSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "name",
          "description",
          "stages",
          "trigger"
        ],
        "title": "CycleDetailResponse",
        "type": "object"
      },
      "CycleExecutionAcceptedResponse": {
        "properties": {
          "events_url": {
            "title": "Events Url",
            "type": "string"
          },
          "execution_id": {
            "title": "Execution Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary_url": {
            "title": "Summary Url",
            "type": "string"
          }
        },
        "required": [
          "execution_id",
          "status",
          "events_url",
          "summary_url"
        ],
        "title": "CycleExecutionAcceptedResponse",
        "type": "object"
      },
...
      "CycleExecutionStatusResponse": {
        "properties": {
          "artifacts_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts Root"
          },
          "cycle": {
            "title": "Cycle",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "events_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Events Path"
          },
          "execution_id": {
            "title": "Execution Id",
            "type": "string"
          },
          "plan_result_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Result Path"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "execution_id",
          "cycle",
          "status",
          "artifacts_root",
          "events_path",
          "plan_result_path",
          "error"
        ],
        "title": "CycleExecutionStatusResponse",
        "type": "object"
      },
      "CycleListResponse": {
        "properties": {
          "config_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Path"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CycleSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "config_path"
        ],
        "title": "CycleListResponse",
        "type": "object"
      },
...
      "CycleTriggerSummary": {
        "properties": {
          "paths": {
            "items": {
              "type": "string"
            },
            "title": "Paths",
            "type": "array"
          },
          "since_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since Ref"
          }
        },
        "required": [
          "paths",
          "since_ref"
        ],
        "title": "CycleTriggerSummary",
        "type": "object"
      },
...
      "GenerateAiSummaryRequest": {
        "properties": {
          "force_refresh": {
            "default": false,
            "title": "Force Refresh",
            "type": "boolean"
          }
        },
        "title": "GenerateAiSummaryRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Testosterone API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/adhoc-executions": {
      "post": {
        "description": "Run one framework directly as a one-stage ``adhoc`` cycle.\n\nSame engine, persistence and event stream as a named cycle: progress and\nstatus live under ``/cycle-executions/{execution_id}``.",
        "operationId": "create_adhoc_execution_api_v1_adhoc_executions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdhocExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleExecutionAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Adhoc Execution",
        "tags": [
          "cycles"
        ]
      }
    },
    "/api/v1/ai/config": {
      "put": {
        "operationId": "update_ai_config_api_v1_ai_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiConfigStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Ai Config",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/api/v1/cycle-executions/{execution_id}": {
      "get": {
        "operationId": "get_cycle_execution_api_v1_cycle_executions__execution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "title": "Execution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleExecutionStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cycle Execution",
        "tags": [
          "cycles"
        ]
      }
    },
    "/api/v1/cycle-executions/{execution_id}/events": {
      "get": {
        "operationId": "stream_cycle_execution_events_api_v1_cycle_executions__execution_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "title": "Execution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Cycle Execution Events",
        "tags": [
          "cycles"
        ]
      }
    },
    "/api/v1/cycles": {
      "get": {
        "description": "List every cycle defined in the resolved configuration (UI parity with ``testo cycles list``).",
        "operationId": "list_cycles_api_v1_cycles_get",
        "parameters": [
          {
            "in": "query",
            "name": "config_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Config Path"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Cycles",
        "tags": [
          "cycles"
        ]
      }
    },
    "/api/v1/cycles/{cycle}": {
      "get": {
        "description": "Resolved view of one cycle (UI parity with ``testo cycles show <name>``).",
        "operationId": "get_cycle_api_v1_cycles__cycle__get",
        "parameters": [
          {
            "in": "path",
            "name": "cycle",
            "required": true,
            "schema": {
              "title": "Cycle",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "config_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Config Path"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cycle",
        "tags": [
          "cycles"
        ]
      }
    },
    "/api/v1/cycles/{cycle}/executions": {
      "post": {
        "operationId": "create_cycle_execution_api_v1_cycles__cycle__executions_post",
        "parameters": [
          {
            "in": "path",
            "name": "cycle",
            "required": true,
            "schema": {
              "title": "Cycle",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CycleExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleExecutionAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Cycle Execution",
        "tags": [
          "cycles"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/ai-summary": {
      "get": {
        "operationId": "get_run_ai_summary_api_v1_runs__run_id__ai_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Ai Summary",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/runs/{run_id}/ai-summary:generate": {
      "post": {
        "operationId": "generate_run_ai_summary_api_v1_runs__run_id__ai_summary_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateAiSummaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Run Ai Summary",
        "tags": [
          "ai"
        ]
      }
    },
...
  }
}